enable username password request everywhere but ios/android
